Vulnerability exposed

+++ os/net/mac/ble/ble-l2cap.c
@@ -105,7 +105,7 @@ get_channel_for_addr(const linkaddr_t *peer_addr)
 l2cap_channel_t *
 get_channel_for_cid(uint16_t own_cid)
 {
-  int16_t i = own_cid - L2CAP_FLOW_CHANNEL;
+  int8_t i = own_cid - L2CAP_FLOW_CHANNEL;
   if(i >= 0 && i < l2cap_channel_count) {
     return &l2cap_channels[own_cid - L2CAP_FLOW_CHANNEL];
   } else {
